# src/etapi/etapi.openapi.yaml
openapi: "3.0.3"
info:
    version: 1.0.0
    title: ETAPI
    description: External Trilium API
    contact:
        name: zadam
        email: zadam.apps@gmail.com
        url: https://github.com/zadam/trilium
    license:
        name: Apache 2.0
        url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
    - url: http://localhost:37740/etapi
    - url: http://localhost:8080/etapi
security:
    - EtapiTokenAuth: []
    - EtapiBasicAuth: []
paths:
...
    /attachments/{attachmentId}:
        parameters:
            - name: attachmentId
              in: path
              required: true
              schema:
                  $ref: "#/components/schemas/EntityId"
        get:
            description: Returns an attachment identified by its ID
            operationId: getAttachmentById
            responses:
                "200":
                    description: attachment response
                    content:
                        application/json; charset=utf-8:
                            schema:
                                $ref: "#/components/schemas/Attachment"
                default:
                    description: unexpected error
                    content:
                        application/json; charset=utf-8:
                            schema:
                                $ref: "#/components/schemas/Error"
        patch:
            description: patch an attachment identified by the attachmentId with changes in the body. Only role, mime, title, and position are patchable.
            operationId: patchAttachmentById
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/Attachment"
            responses:
                "200":
                    description: attribute updated
                    content:
                        application/json; charset=utf-8:
                            schema:
                                $ref: "#/components/schemas/Attachment"
                default:
                    description: unexpected error
                    content:
                        application/json; charset=utf-8:
                            schema:
                                $ref: "#/components/schemas/Error"
        delete:
            description: deletes an attachment based on the attachmentId supplied.
            operationId: deleteAttachmentById
            responses:
                "204":
                    description: attachment deleted
                default:
                    description: unexpected error
                    content:
                        application/json; charset=utf-8:
                            schema:
                                $ref: "#/components/schemas/Error"
...
components:
    securitySchemes:
        EtapiTokenAuth:
            type: apiKey
            in: header
            name: Authorization
        EtapiBasicAuth:
            type: http
            scheme: basic
            description: >
                Basic Auth where username is arbitrary string (e.g. "trilium", not checked), 
                username is the ETAPI token. 
                To emphasize, do not use Trilium password here (won't work), only the generated
                ETAPI token (from Options -> ETAPI)
    schemas:
...
        Attachment:
            type: object
            description: Attachment is owned by a note, has title and content
            properties:
                attachmentId:
                    $ref: "#/components/schemas/EntityId"
                    readOnly: true
                ownerId:
                    $ref: "#/components/schemas/EntityId"
                    description: identifies the owner of the attachment, is either noteId or revisionId
                role:
                    type: string
                mime:
                    type: string
                title:
                    type: string
                position:
                    type: integer
                    format: int32
                blobId:
                    type: string
                    description: ID of the blob object which effectively serves as a content hash
                dateModified:
                    $ref: "#/components/schemas/LocalDateTime"
                    readOnly: true
                utcDateModified:
                    $ref: "#/components/schemas/UtcDateTime"
                    readOnly: true
                utcDateScheduledForErasureSince:
                    $ref: "#/components/schemas/UtcDateTime"
                    readOnly: true
                contentLength:
                    type: integer
                    format: int32
...
        EntityId:
            type: string
            pattern: "[a-zA-Z0-9_]{4,32}"
            example: evnnmvHTCgIn
...
        LocalDateTime:
            type: string
            pattern: '[0-9]{4}-[0-9]{2}-[0-9]{2} [0-9]{2}:[0-9]{2}:[0-9]{2}\.[0-9]{3}[\+\-][0-9]{4}'
            example: 2021-12-31 20:18:11.930+0100
        UtcDateTime:
            type: string
            pattern: '[0-9]{4}-[0-9]{2}-[0-9]{2} [0-9]{2}:[0-9]{2}:[0-9]{2}\.[0-9]{3}Z'
            example: 2021-12-31 19:18:11.930Z
...
        Error:
            type: object
            required:
                - status
                - code
                - message
            properties:
                status:
                    type: integer
                    format: int32
                    description: HTTP status, identical to the one given in HTTP response
                    example: 400
                code:
                    type: string
                    description: stable string constant
                    example: NOTE_IS_PROTECTED
                message:
                    type: string
                    description: Human readable error, potentially with more details,
                    example: Note 'evnnmvHTCgIn' is protected and cannot be modified through ETAPI